Fixed Perfomance Regression for Shrinking Window Frames (+Rebase)
